refactor: move models to internal package
