+++ nebula/components/home/sections/hero.templ
@@ -1,8 +1,8 @@
 package sections
 
 import (
+	"github.com/onsonr/sonr/internal/orm/models"
 	"github.com/onsonr/sonr/nebula/global/ui"
-	"github.com/onsonr/sonr/nebula/models"
 )
 
 templ Hero(hero *models.Hero) {
