+++ nebula/components/home/sections/highlights.templ
@@ -1,6 +1,6 @@
 package sections
 
-import "github.com/onsonr/sonr/nebula/models"
+import "github.com/onsonr/sonr/internal/orm/models"
 
 templ Highlights(highlights *models.Highlights) {
 	<!-- Features -->
